pyproject.toml
[project]
name = "llm-agents-mcp-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai-tools>=1.6.1",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "langchain==1.2.12",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.1",
    "mcp[cli]>=1.27.0",
    "mlflow[genai,langchain]>=3.11.1",
    "python-dotenv>=1.2.2",
]
